requirements.txt
numpy~=1.26.3
scikit-learn~=1.5.2
torch~=2.4.1+cu121
einops~=0.8.0
opencv-python~=4.10.0.84
matplotlib~=3.9.2
pillow~=10.2.0
torchvision~=0.19.1+cu121
scipy~=1.14.1
h5py~=3.12.1